platform_data/mlxreg: Extend hotplug item structure with kind field
And new field 'kind' to 'mlxreg_core_item' structure.
This field is of type 'mlxreg_hotplug_kind' and it specifies
the type of hotplug item entry.
The purpose of this filed is to allow appropriate handling of hotplug
events in modular system.
For example line card events, excepting standard type of hardware
events, like insertion/removal or power on/off, also include security
event, which are raised by hardware after line card hardware
verification and validation of security.

Signed-off-by: Vadim Pasternak <vadimp@mellanox.com>