leds: mlxreg: Allow multi-instantiation of same name LED for modular systems
It could be more than one instance of LED with the same name in the
modular systems. For example, "status" or "uid" LED can be located
on chassis and on each line card of modular system.
In order to avoid conflicts with duplicated names, append platform
device Id, which is unique, to LED name after driver name.
Thus, for example, "status" LED on chassis is to be named
"mlxreg:status:green". As it is named now on non-modular systems, on
which platform device Id is not specified. While for the line cards
LEDs it will be called like: "mlxreg48:status:green",
"mlxreg66:status:green", etcetera.

Signed-off-by: Vadim Pasternak <vadimp@mellanox.com>